image_processing/requirements.txt
boto3==1.5.36
Pillow==5.0.0
numpy==1.14.0